+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 
 ### Added
-- Add multi-select support to filter select on the dashboard
+- web: Add multi-select support to filter select on the dashboard
+- gitseek: add StatusLangFileUpToDate constant for up-to-date language files
 
 ## [v0.0.0] - 2026-03-14